Reorganize shader module into smaller files
