+++ vispy/visuals/shaders/function.py
@@ -21,166 +21,14 @@ import re
 import logging
 import numpy as np
 
-from ...util.event import EventEmitter, Event
 from ...util.eq import eq
 from ...util import logger
 from ...ext.ordereddict import OrderedDict
 from ...ext.six import string_types
 from . import parsing
-from .compiler import Compiler
-
-VARIABLE_TYPES = ('const', 'uniform', 'attribute', 'varying', 'inout')
-
-
-class ShaderChangeEvent(Event):
-    def __init__(self, code_changed=False, value_changed=False, **kwargs):
-        Event.__init__(self, type='shader_change', **kwargs)
-        self.code_changed = code_changed
-        self.value_changed = value_changed
-
-
-class ShaderObject(object):
-    """ Base class for all objects that may be included in a GLSL program
-    (Functions, Variables, Expressions).
-    
-    Shader objects have a *definition* that defines the object in GLSL, an 
-    *expression* that is used to reference the object, and a set of 
-    *dependencies* that must be declared before the object is used.
-    
-    Dependencies are tracked hierarchically such that changes to any object
-    will be propagated up the dependency hierarchy to trigger a recompile.
-    """
-    
-    @classmethod
-    def create(self, obj, ref=None):
-        """ Convert *obj* to a new ShaderObject. If the output is a Variable
-        with no name, then set its name using *ref*. 
-        """
-        if isinstance(ref, Variable):
-            ref = ref.name
-        elif isinstance(ref, string_types) and ref.startswith('gl_'):
-            # gl_ names not allowed for variables
-            ref = ref[3:].lower()
-        
-        # Allow any type of object to be converted to ShaderObject if it
-        # provides a magic method:
-        if hasattr(obj, '_shader_object'):
-            obj = obj._shader_object()
-        
-        if isinstance(obj, ShaderObject):
-            if isinstance(obj, Variable) and obj.name is None:
-                obj.name = ref
-        elif isinstance(obj, string_types):
-            obj = TextExpression(obj)
-        else:
-            obj = Variable(ref, obj)
-            # Try prepending the name to indicate attribute, uniform, varying
-            if obj.vtype and obj.vtype[0] in 'auv':
-                obj.name = obj.vtype[0] + '_' + obj.name 
-        
-        return obj
-    
-    def __init__(self):
-        # emitted when any part of the code for this object has changed,
-        # including dependencies.
-        self.changed = EventEmitter(source=self, event_class=ShaderChangeEvent)
-        
-        # objects that must be declared before this object's definition.
-        # {obj: refcount}
-        self._deps = OrderedDict()  # OrderedDict for consistent code output
-    
-    @property
-    def name(self):
-        """ The name of this shader object.
-        """
-        return None
-        
-    def definition(self, obj_names):
-        """ Return the GLSL definition for this object. Use *obj_names* to
-        determine the names of dependencies.
-        """
-        return None
-    
-    def expression(self, obj_names):
-        """ Return the GLSL expression used to reference this object inline.
-        """
-        return obj_names[self]
-    
-    def dependencies(self, sort=False):
-        """ Return all dependencies required to use this object. The last item 
-        in the list is *self*.
-        """
-        alldeps = []
-        if sort:
-            def key(obj):
-                # sort deps such that we get functions, variables, self.
-                if not isinstance(obj, Variable):
-                    return (0, 0)
-                else:
-                    return (1, obj.vtype)
-            
-            deps = sorted(self._deps, key=key)
-        else:
-            deps = self._deps
-        
-        for dep in deps:
-            alldeps.extend(dep.dependencies(sort=sort))
-        alldeps.append(self)
-        return alldeps
-
-    def static_names(self):
-        """ Return a list of names that are declared in this object's 
-        definition (not including the name of the object itself).
-        
-        These names will be reserved by the compiler when automatically 
-        determining object names.
-        """
-        return []
-    
-    def _add_dep(self, dep):
-        """ Increment the reference count for *dep*. If this is a new 
-        dependency, then connect to its *changed* event.
-        """
-        if dep in self._deps:
-            self._deps[dep] += 1
-        else:
-            self._deps[dep] = 1
-            dep.changed.connect(self._dep_changed)
-
-    def _remove_dep(self, dep):
-        """ Decrement the reference count for *dep*. If the reference count 
-        reaches 0, then the dependency is removed and its *changed* event is
-        disconnected.
-        """
-        refcount = self._deps[dep]
-        if refcount == 1:
-            self._deps.pop(dep)
-            dep.changed.disconnect(self._dep_changed)
-        else:
-            self._deps[dep] -= 1
-        
-    def _dep_changed(self, event):
-        """ Called when a dependency's expression has changed.
-        """
-        logger.debug("ShaderObject changed: %r" % event.source)
-        self.changed(event)
-    
-    def compile(self):
-        """ Return a compilation of this object and its dependencies. 
-        
-        Note: this is mainly for debugging purposes; the names in this code
-        are not guaranteed to match names in any other compilations. Use
-        Compiler directly to ensure consistent naming across multiple objects. 
-        """
-        compiler = Compiler(obj=self)
-        return compiler.compile()['obj']
-    
-    def __repr__(self):
-        if self.name is not None:
-            return '<%s "%s" at 0x%x>' % (self.__class__.__name__, 
-                                          self.name, id(self))
-        else:
-            return '<%s at 0x%x>' % (self.__class__.__name__, id(self))
+from .shader_object import ShaderObject
+from .variable import Variable, Varying
+from .expression import TextExpression, FunctionCall
 
 
 class Function(ShaderObject):
@@ -695,306 +543,6 @@ class MainFunction(Function):
     
     def remove_callback(self, hook, func):
         self._chains[hook].remove(func)
-
-
-class Variable(ShaderObject):
-    """ Representation of global shader variable
-    
-    Parameters
-    ----------
-    name : str
-        the name of the variable. This string can also contain the full
-        definition of the variable, e.g. 'uniform vec2 foo'.
-    value : {float, int, tuple, GLObject}
-        If given, vtype and dtype are determined automatically. If a
-        float/int/tuple is given, the variable is a uniform. If a gloo
-        object is given that has a glsl_type property, the variable is
-        an attribute and
-    vtype : {'const', 'uniform', 'attribute', 'varying', 'inout'}
-        The type of variable.
-    dtype : str
-        The data type of the variable, e.g. 'float', 'vec4', 'mat', etc.
-    
-    """
-    def __init__(self, name, value=None, vtype=None, dtype=None):
-        super(Variable, self).__init__()
-        
-        # allow full definition in first argument
-        if ' ' in name:
-            fields = name.split(' ')
-            if len(fields) == 3:
-                vtype, dtype, name = fields
-            elif len(fields) == 4 and fields[0] == 'const':
-                vtype, dtype, name, value = fields
-            else:
-                raise ValueError('Variable specifications given by string must'
-                                 ' be of the form "vtype dtype name" or '
-                                 '"const dtype name value".')
-            
-        if not (isinstance(name, string_types) or name is None):
-            raise TypeError("Variable name must be string or None.")
-        
-        self._state_counter = 0
-        self._name = name
-        self._vtype = vtype
-        self._dtype = dtype
-        self._value = None
-        
-        # If vtype/dtype were given at init, then we will never
-        # try to set these values automatically.
-        self._type_locked = self._vtype is not None and self._dtype is not None
-            
-        if value is not None:
-            self.value = value
-        
-        if self._vtype and self._vtype not in VARIABLE_TYPES:
-            raise ValueError('Not a valid vtype: %r' % self._vtype)
-    
-    @property
-    def name(self):
-        """ The name of this variable.
-        """
-        return self._name
-    
-    @name.setter
-    def name(self, n):
-        # Settable mostly to allow automatic setting of varying names 
-        # See ShaderObject.create()
-        if self._name != n:
-            self._name = n
-            self.changed(code_changed=True)
-    
-    @property
-    def vtype(self):
-        """ The type of variable (const, uniform, attribute, varying or inout).
-        """
-        return self._vtype
-    
-    @property
-    def dtype(self):
-        """ The type of data (float, int, vec, mat, ...).
-        """
-        return self._dtype
-    
-    @property
-    def value(self):
-        """ The value associated with this variable.
-        """
-        return self._value
-    
-    @value.setter
-    def value(self, value):
-        if isinstance(value, (tuple, list)) and 1 < len(value) < 5:
-            vtype = 'uniform'
-            dtype = 'vec%d' % len(value)
-        elif isinstance(value, np.ndarray):
-            if value.ndim == 1 and (1 < len(value) < 5):
-                vtype = 'uniform'
-                dtype = 'vec%d' % len(value)
-            elif value.ndim == 2 and value.shape in ((2, 2), (3, 3), (4, 4)):
-                vtype = 'uniform'
-                dtype = 'mat%d' % value.shape[0]                
-            else:
-                raise ValueError("Cannot make uniform value for %s from array "
-                                 "of shape %s." % (self.name, value.shape))
-        elif np.isscalar(value):
-            vtype = 'uniform'
-            if isinstance(value, (float, np.floating)):
-                dtype = 'float'
-            elif isinstance(value, (int, np.integer)):
-                dtype = 'int'
-            else:
-                raise TypeError("Unknown data type %r for variable %r" % 
-                                (type(value), self))
-        elif getattr(value, 'glsl_type', None) is not None:
-            # Note: hasattr() is broken by design--swallows all exceptions!
-            vtype, dtype = value.glsl_type
-        else:
-            raise TypeError("Unknown data type %r for variable %r" % 
-                            (type(value), self))
-
-        self._value = value
-        self._state_counter += 1
-        
-        if self._type_locked:
-            if dtype != self._dtype or vtype != self._vtype:
-                raise TypeError('Variable is type "%s"; cannot assign value '
-                                '%r.' % (self.dtype, value))
-            return
-            
-        # update vtype/dtype and emit changed event if necessary
-        changed = False
-        if self._dtype != dtype:
-            self._dtype = dtype
-            changed = True
-        if self._vtype != vtype:
-            self._vtype = vtype
-            changed = True
-        if changed:
-            self.changed(code_changed=True, value_changed=True)
-    
-    @property
-    def state_id(self):
-        """Return a unique ID that changes whenever the state of the Variable
-        has changed. This allows ModularProgram to quickly determine whether
-        the value has changed since it was last used."""
-        return id(self), self._state_counter
-
-    def __repr__(self):
-        return ("<%s \"%s %s %s\" at 0x%x>" % (self.__class__.__name__,
-                                               self._vtype, self._dtype, 
-                                               self.name, id(self)))
-    
-    def expression(self, names):
-        return names[self]
-    
-    def definition(self, names):
-        if self.vtype is None:
-            raise RuntimeError("Variable has no vtype: %r" % self)
-        if self.dtype is None:
-            raise RuntimeError("Variable has no dtype: %r" % self)
-        
-        name = names[self]
-        if self.vtype == 'const':
-            return '%s %s %s = %s;' % (self.vtype, self.dtype, name, 
-                                       self.value)
-        else:
-            return '%s %s %s;' % (self.vtype, self.dtype, name)
-
-
-class Varying(Variable):
-    """ Representation of a varying
-    
-    Varyings can inherit their dtype from another Variable, allowing for
-    more flexibility in composing shaders.
-    """
-    def __init__(self, name, dtype=None):
-        self._link = None
-        Variable.__init__(self, name, vtype='varying', dtype=dtype)
-        
-    @property
-    def value(self):
-        """ The value associated with this variable.
-        """
-        return self._value
-    
-    @value.setter
-    def value(self, value):
-        if value is not None:
-            raise TypeError("Cannot assign value directly to varying.")
-    
-    @property
-    def dtype(self):
-        if self._dtype is None:
-            if self._link is None:
-                return None
-            else:
-                return self._link.dtype
-        else:
-            return self._dtype
-
-    def link(self, var):
-        """ Link this Varying to another object from which it will derive its
-        dtype. This method is used internally when assigning an attribute to
-        a varying using syntax ``aFunction[varying] = attr``.
-        """
-        assert self._dtype is not None or hasattr(var, 'dtype')
-        self._link = var
-        self.changed()
-
-
-class Expression(ShaderObject):
-    """ Base class for expressions (ShaderObjects that do not have a
-    definition nor dependencies)
-    """
-    
-    def definition(self, names):
-        # expressions are declared inline.
-        return None
-
-
-class TextExpression(Expression):
-    """ Plain GLSL text to insert inline
-    """
-    
-    def __init__(self, text):
-        super(TextExpression, self).__init__()
-        if not isinstance(text, string_types):
-            raise TypeError("Argument must be string.")
-        self._text = text
-    
-    def __repr__(self):
-        return '<TextExpression %r for at 0x%x>' % (self.text, id(self))
-    
-    def expression(self, names=None):
-        return self._text
-    
-    @property
-    def text(self):
-        return self._text
-    
-    @text.setter
-    def text(self, t):
-        self._text = t
-        self.changed()
-
-    def __eq__(self, a):
-        if isinstance(a, TextExpression):
-            return a._text == self._text
-        elif isinstance(a, string_types):
-            return a == self._text
-        else:
-            return False
-    
-    def __hash__(self):
-        return self._text.__hash__()
-
-
-class FunctionCall(Expression):
-    """ Representation of a call to a function
-    
-    Essentially this is container for a Function along with its signature. 
-    """
-    def __init__(self, function, args):
-        super(FunctionCall, self).__init__()
-        
-        if not isinstance(function, Function):
-            raise TypeError('FunctionCall needs a Function')
-        
-        sig_len = len(function.args)
-        if len(args) != sig_len:
-            raise TypeError('Function %s requires %d arguments (got %d)' %
-                            (function.name, sig_len, len(args)))
-        
-        # Ensure all expressions
-        sig = function.args
-        
-        self._function = function
-        
-        # Convert all arguments to ShaderObject, using arg name if possible.
-        self._args = [ShaderObject.create(arg, ref=sig[i][1]) 
-                      for i, arg in enumerate(args)]
-        
-        self._add_dep(function)
-        for arg in self._args:
-            self._add_dep(arg)
-    
-    def __repr__(self):
-        return '<FunctionCall of %r at 0x%x>' % (self.function.name, id(self))
-    
-    @property
-    def function(self):
-        return self._function
-    
-    @property
-    def dtype(self):
-        return self._function.rtype
-    
-    def expression(self, names):
-        str_args = [arg.expression(names) for arg in self._args]
-        args = ', '.join(str_args)
-        fname = self.function.expression(names)
-        return '%s(%s)' % (fname, args)
 
 
 class FunctionChain(Function):
